adding equals and hashcode to Subcontractor and changed implementation of setCustomerId
Override of equals method to do comparrison on name and city only.
There are no unique identifiers. A limitation of the Subcontractor class

The setCustomerId implementation has been changed to use a String to
handle edge case where customerId is 00000001.
Tests added to cover the test case.

Override of hashcode method. Again this is using a hash of the concat of
the subcontractor name and city.